client code refactoring
