+++ client/src/components/App.js
@@ -1,12 +1,25 @@
+// Copyright (c) 2019 Cisco and/or its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import React from 'react'
-import PluginApp from './Plugin/PluginApp';
-import GeneratorApp from './Generator/GeneratorApp';
-//import ProjectSelection from './Project-Selection/ProjectSelection';
 import { BrowserRouter, Route } from 'react-router-dom';
 
+import PluginApp from './PluginApp/PluginApp';
+import GeneratorApp from './GeneratorApp/GeneratorApp';
+
 /*
-* This defines the three routes from the web app, currently there
-* is the default page (project selection), the plugin app and
+* This defines the three routes from the web app, the plugin app and
 * the generator app.
 */
 class App extends React.Component {
@@ -24,5 +37,3 @@ class App extends React.Component {
   }
 }
 export default App;
-
-
