--- client/src/components/Generator/TreeNode.js
@@ -1,88 +1,0 @@
-import React from 'react';
-import { FaFile, FaFolder, FaFolderOpen, FaChevronDown, FaChevronRight } from 'react-icons/fa';
-import styled from 'styled-components';
-import last from 'lodash/last';
-import PropTypes from 'prop-types';
-import "../../styles_CSS/Generator/GeneratorApp.css";
-
-let i = 0;
-
-const getPaddingLeft = (level, type) => {
-  let paddingLeft = level * 20;
-  if (type === 'file') paddingLeft += 20;
-  return paddingLeft;
-}
-
-const StyledTreeNode = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding: 5px 20px;
-  cursor: pointer;
-  -webkit-user-select: none; /* Safari */        
-  -moz-user-select: none; /* Firefox */
-  -ms-user-select: none; /* IE10+/Edge */
-  user-select: none; /* Standard */
-  padding-left: ${props => getPaddingLeft(props.level, props.type)}px;
-  &:hover {
-    background: lightgray;
-  }
-`;
-
-const NodeIcon = styled.div`
-  font-size: 12px;
-  margin-right: ${props => props.marginRight ? props.marginRight : 5}px;
-  color: #e8bb4d;
-`;
-
-const getNodeLabel = (node) => last(node.path.split('/'));
-
-
-
-const TreeNode = (props) => {
-  const { node, getChildNodes, level, onToggle } = props;
-  return (
-    <React.Fragment>
-      <StyledTreeNode level={level} type={node.type} onClick={() => onToggle(node)}>
-      <NodeIcon >
-      { node.type === 'folder' && (node.isOpen ? <FaChevronDown /> : <FaChevronRight />) }
-        
-        </NodeIcon>
-        
-        <NodeIcon marginRight={10}>
-          { node.type === 'file' && <FaFile /> }
-          { node.type === 'folder' && node.isOpen === true && <FaFolderOpen /> }
-          { node.type === 'folder' && !node.isOpen && <FaFolder /> }
-        </NodeIcon>
-        
-
-        <span className="node-name">
-          { getNodeLabel(node) }
-        </span>
-      </StyledTreeNode>
-
-      { node.isOpen && getChildNodes(node).map(childNode => (
-        <TreeNode 
-          {...props}
-          node={childNode}
-          key={i++}          
-          level={level + 1}
-        />
-      ))}
-    </React.Fragment>
-  );
-}
-
-TreeNode.propTypes = {
-  node: PropTypes.object.isRequired,
-  getChildNodes: PropTypes.func.isRequired,
-  level: PropTypes.number.isRequired,
-  onToggle: PropTypes.func.isRequired,
-  onNodeSelect: PropTypes.func.isRequired,
-};
-
-TreeNode.defaultProps = {
-  level: 0,
-};
-
-export default TreeNode;
