+++ client/src/components/PluginApp/PluginPalette.js
@@ -1,37 +1,59 @@
+// Copyright (c) 2019 Cisco and/or its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import React from 'react';
 import PropTypes from 'prop-types'
 import Swal from 'sweetalert2'
+
 import "../../styles_CSS/App.css";
 import "../../styles_CSS/Plugin/Splitright.css";
 
-
+// PluginPalette globals
 let pluginModule = require('../Model');
-
+let textVisibilty = 'visible';
 const toolText = {
   generate: 'Once finished, generate a template.',
   header: 'Configure and save projects.',
   pluginPicker: 'Pick from a set of plugins or make your own.',
   agent: 'Choose agent settings.'
 }
-let textVisibilty = 'visible';
 
-/*
-* This component represents the main workspace. Users will be able to 
-* Drag and Drop into the area that this component represents.
-*/
+/***************************************************************
+* This component defines the logic of how the plugin picker and 
+* the plugin palette interface as well as which selectable
+* plugins are within which view.
+* 
+* PluginApp.js --> PluginPicker.js    --> SelectablePlugins.js
+*              --> PluginPalette.js   --> SelectablePlugins.js
+*              --> PluginAppHeader.js --> Dropdown.js
+****************************************************************/
+
 const PluginPalette = (props) => {
   var pluginArray = React.Children.toArray(props.children);
-  //This will loop through pluginArray splicing out elements based
-  //on if the element of sentInArrayObject at the index of the 
-  //current counter (i) is === 0. The net result is that if for example 
-  //sentInArrayObject = [0,1,1,0], plugins with the id attribute 1 and 2
-  //are rendered within PluginPalette and not PluginPicker.
+  
+  // Loop through to find out which plugins are not in the plugin picker. 1 = in plugin picker.
   for(let i = props.sentInArray.length; i >= 0; i--) {
     if(props.sentInArray[i] === 0 || props.sentInArray[i] === false) { pluginArray.splice(i,1); }
   }
   if(props.sentInArray.includes(1)) textVisibilty = 'hidden';
   else textVisibilty = 'visible';
 
+  /*
+  ================================
+  Render
+  ================================
+  */
   return (
     <div>
       <div >
@@ -41,6 +63,7 @@ const PluginPalette = (props) => {
               {pluginArray}
             </div>
           </div>
+          {/* Renders the help text if there are no plugins selected */}
           <div className="tool-text-container" style={{visibility: textVisibilty}}>
             <p className="tool-text-header">{toolText.generate}&emsp;<i className="up-arrow"></i></p>
             <div className="tool-text-div">
@@ -51,6 +74,7 @@ const PluginPalette = (props) => {
           </div>
         </div>
       </div>
+      {/* Renders the bottom agent box along with it configuration button */}
       <div className="split right-bottom">
         <div className="rectangle">
           <img
@@ -64,6 +88,8 @@ const PluginPalette = (props) => {
             <p className="agent-name">Agent</p>
           </div>
         </div>
+        {/* Renders the cisco logo if the window width is less than 1400px
+        this logic exists in Splitright.css */}
         <img
           className="cisco-logo-gray"
           src='/images/cisco-logo.png'
@@ -79,6 +105,12 @@ PluginPalette.propTypes = {
   sentInArray:   PropTypes.array.isRequired,
 }
 
+/*
+================================
+Helper Functions
+================================
+*/
+// Shows an agent name configuration using a swal2 (popup library) popup.
 function handleNewAgentName() {
   (async () => {
     let nameCapture = await getAgentName();
@@ -88,6 +120,7 @@ function handleNewAgentName() {
   })()
 }
 
+// Return the agent name from the popup
 async function getAgentName () {
   const inputStyling = `<div style="display:inline-block; width: 300px;">
                           <p style="float: left; width: 100px; margin-bottom: -15px;">Agent Name:</p>
