+++ client/src/redux/actions/index.js
@@ -1,3 +1,17 @@
+// Copyright (c) 2019 Cisco and/or its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 // Used by UI
 import { ADD_CURR_PROJECT } from "../constants/action-types";
 import { SET_CURR_PROJECT } from "../constants/action-types";
@@ -49,4 +63,3 @@ export function returnLoadProject(payload) {
 export function returnTemplate(payload) {
   return { type: RETURN_TEMPLATE, payload }; 
 }
-
