+++ client/src/redux/index.js
@@ -1,3 +1,17 @@
+// Copyright (c) 2019 Cisco and/or its affiliates.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import store from "../redux/store/index";
 // Used by UI
 import { addCurrProject } from "../redux/actions/index";
@@ -29,4 +43,3 @@ window.downloadTar = downloadTar;
 
 window.returnLoadProject = returnLoadProject;
 window.returnTemplate = returnTemplate;
-
